moving to the counter file
