+++ flutter_mobx/example/lib/counter.dart
@@ -1,0 +1,54 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:mobx/mobx.dart';
+
+class CounterExample extends StatefulWidget {
+  const CounterExample({Key key, this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  _CounterExampleState createState() => _CounterExampleState();
+}
+
+class _CounterExampleState extends State<CounterExample> {
+  _CounterExampleState() {
+    incrementCounter = action(() {
+      counter++;
+    });
+  }
+
+  final Observable<int> _counter = Observable(0);
+
+  int get counter => _counter.value;
+  set counter(int value) => _counter.value = value;
+
+  Action incrementCounter;
+
+  @override
+  Widget build(BuildContext context) => Scaffold(
+        appBar: AppBar(
+          title: Text(widget.title),
+        ),
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              const Text(
+                'You have pushed the button this many times:',
+              ),
+              Observer(
+                  builder: (_) => Text(
+                        '$counter',
+                        style: const TextStyle(fontSize: 20),
+                      )),
+            ],
+          ),
+        ),
+        floatingActionButton: FloatingActionButton(
+          onPressed: incrementCounter,
+          tooltip: 'Increment',
+          child: const Icon(Icons.add),
+        ),
+      );
+}
