+++ flutter_mobx/example/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mobx/mobx.dart';
-import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:flutter_mobx_example/counter.dart';
 
 void main() => runApp(MyApp());
 
@@ -13,56 +12,5 @@ class MyApp extends StatelessWidget {
           primarySwatch: Colors.blue,
         ),
         home: const CounterExample(title: 'Flutter Demo Home Page'),
-      );
-}
-
-class CounterExample extends StatefulWidget {
-  const CounterExample({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _CounterExampleState createState() => _CounterExampleState();
-}
-
-class _CounterExampleState extends State<CounterExample> {
-  _CounterExampleState() {
-    incrementCounter = action(() {
-      counter++;
-    });
-  }
-
-  final Observable<int> _counter = Observable(0);
-
-  int get counter => _counter.value;
-  set counter(int value) => _counter.value = value;
-
-  Action incrementCounter;
-
-  @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(
-          title: Text(widget.title),
-        ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              const Text(
-                'You have pushed the button this many times:',
-              ),
-              Observer(
-                  builder: (_) => Text(
-                        '$counter',
-                        style: const TextStyle(fontSize: 20),
-                      )),
-            ],
-          ),
-        ),
-        floatingActionButton: FloatingActionButton(
-          onPressed: incrementCounter,
-          tooltip: 'Increment',
-          child: const Icon(Icons.add),
-        ),
       );
 }
